+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Add a new wizard helper for setting Lizmap Web Client groups in the plugin
+* Check the login and server URL before saving in the authentication database
 
 ## 3.9.8 - 2022-12-08